feat/#2 : 메뉴 단건/다건 조회 기능 예외처리
--------------충돌방지 no commit --------------------
entity / Menu,Store @Getter 활성화
repository/StoreRepository  임시생성

+++ src/main/java/like/heocholi/spartaeats/controller/MenuController.java
@@ -17,13 +17,13 @@ public class MenuController {
     // R
     //메뉴 단건 조회
     @GetMapping("/stores/{storeId}/menus/{menuId}")
-    public ResponseEntity<ResponseMessage> getMenu(@PathVariable int storeId, @PathVariable int menuId) {
+    public ResponseEntity<ResponseMessage> getMenu(@PathVariable Long storeId, @PathVariable Long menuId) {
         return menuService.getMenu(storeId,menuId);
     }
 
     //메뉴 전체 조회
     @GetMapping("/stores/{storeId}/menus")
-    public ResponseEntity<ResponseMessage> getMenus(@PathVariable int storeId) {
+    public ResponseEntity<ResponseMessage> getMenus(@PathVariable Long storeId) {
         return menuService.getMenus(storeId);
     }
