+++ src/main/java/like/heocholi/spartaeats/repository/MenuRepository.java
@@ -6,8 +6,8 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import java.util.List;
 import java.util.Optional;
 
-public interface MenuRepository extends JpaRepository<Menu, Integer> {
+public interface MenuRepository extends JpaRepository<Menu, Long> {
 
-    Optional<List<Menu>> findAllByStoreId(int storeId);
-    Optional<Menu> findByStoreIdAndMenuId(int storeId, int menuId);
+    Optional<List<Menu>> findAllByStoreId(Long storeId);
+    Optional<Menu> findByStoreIdAndMenuId(Long storeId, Long menuId);
 }
